feat(link-forever): linking by URLs to remote HEAD (#12)
Add a (currently undocumented) compatibility feature where HTTP links to
HEAD, such as `https://github.com/owner/repo/tree/HEAD/path/to/file.png`,
will be treated as intended to be a permalink and converted/checked
accordingly, as if they were written as a path-based link.

This can be used to, for example, convert image links in included files
(`{{#include}}`) back to path-based links, so that they become
same-origin links in book output while staying as GitHub links when
viewed elsewhere.